modify memmove

+++ c01Libft/ft_memset.c
@@ -1,8 +1,8 @@
 #include "libft.h"
 
-void	*ft_memset(void *b, int c, int len)
+void	*ft_memset(void *b, int c, size_t len)
 {
-	int				i;
+	size_t			i;
 	unsigned char	*p;
 
 	p = (unsigned char *)b;
